feat: Introduce flagd kube proxy (#495)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
<!-- add the description of the PR here -->

- introduced flagd Kubernetes proxy as new binary 

requires testing and documentation

### Related Issues
<!-- add here the GitHub issue that this PR resolves if applicable -->


### Notes
<!-- any additional notes for this PR -->

### Follow-up Tasks
<!-- anything that is related to this PR but not done here should be
noted under this section -->
<!-- if there is a need for a new issue, please link it here -->
create issues for certificate handling
https://github.com/open-feature/flagd/issues/524

### How to test
<!-- if applicable, add testing instructions under this section -->

Resolves: https://github.com/open-feature/flagd/issues/500

---------

+++ flagd/cmd/start.go
@@ -142,10 +142,10 @@ var startCmd = &cobra.Command{
 		// Build Runtime -----------------------------------------------------------
 		rt, err := runtime.FromConfig(logger, runtime.Config{
 			CORS:              viper.GetStringSlice(corsFlagName),
-			MetricsPort:       viper.GetInt32(metricsPortFlagName),
+			MetricsPort:       viper.GetUint16(metricsPortFlagName),
 			ServiceCertPath:   viper.GetString(serverCertPathFlagName),
 			ServiceKeyPath:    viper.GetString(serverKeyPathFlagName),
-			ServicePort:       viper.GetInt32(portFlagName),
+			ServicePort:       viper.GetUint16(portFlagName),
 			ServiceSocketPath: viper.GetString(socketPathFlagName),
 			SyncProviders:     syncProviders,
 		})
